+++ src/modules/calendar/index.ts
@@ -1,0 +1,2 @@
+export * from './calendar.types';
+export { default as CalendarPage } from './CalendarPage';